split aot test utils
